+++ Sources/CardPayments/Models/OrdersV2/PaymentSource.swift
@@ -1,6 +1,6 @@
 import Foundation
-#if canImport(PaymentsCore)
-import PaymentsCore
+#if canImport(CorePayments)
+import CorePayments
 #endif
 
 public struct PaymentSource: Decodable {